datacube config path is now taken from the django settings
Previously it was hardcoded to be `/app/config/datacube.conf` and
it was also being stored as a class attribute in most apps task classes.
It is now taken from the `DATACUBE_CONFIG_PATH` django setting, which in
turn is taken from the environment.

Also added a custom data access api wrapper class that allows specifying
an environment name